refactor(weather): Update ViewModel provider imports in screen widgets
- Modified import paths in error_display_screen.dart and weather_result_screen.dart
to reference error and city weather providers from features/weather/view_model/.
